api: prepare for enabling async io
Reorganize filesystem, server, vfs and passthroughfs to prepare for
enabling rust asynchronous io.
